consolidate controllers, cleanup tests
