+++ open-health-manager-app/src/main/java/org/mitre/healthmanager/repository/FHIRClientRepository.java
@@ -3,10 +3,8 @@ package org.mitre.healthmanager.repository;
 import java.util.List;
 
 import org.mitre.healthmanager.domain.FHIRClient;
-import org.springframework.data.jpa.repository.*;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import org.springframework.data.repository.query.Param;
-import java.util.List;
 
 /**
  * Spring Data SQL repository for the FHIRClient entity.
@@ -15,7 +13,4 @@ import java.util.List;
 public interface FHIRClientRepository extends JpaRepository<FHIRClient, Long> {
 	
     List<FHIRClient> findByFhirOrganizationId(String fhirOrganizationId);
-    
-    @Query("select fHIRClient from FHIRClient fHIRClient where fhirOrganizationId =:id")
-    List<FHIRClient> findAllForFhirOrganizationID(@Param("id") String id);
 }
